+++ R/get_estimates.R
@@ -10,7 +10,7 @@
 #' @param model a single model object
 #' 
 #' @export
-get_estimates <- function(model, conf_level = .95, vcov = NULL, ...) {
+get_estimates <- function(model, conf_level = .95, vcov = NULL, shape = NULL, ...) {
 
     if (is.null(conf_level)) {
         conf_int <- FALSE
@@ -111,7 +111,7 @@ These errors messages were generated during extraction:
     }
 
     # fixest mods
-    fixest_mod = inherits(model, "fixest") || inherits(model, "fixest_multi")
+    fixest_mod <- inherits(model, "fixest") || inherits(model, "fixest_multi")
 
     # vcov override
     flag1 <- !is.null(vcov)
@@ -142,9 +142,14 @@ These errors messages were generated during extraction:
       }
     }
 
+    # combine columns if requested in `shape` argument using an : interaction
+    for (x in shape$combine) {
+        vars <- strsplit(x, ":")[[1]]
+        out[[vars[1]]] <- paste(out[[vars[1]]], out[[vars[2]]])
+    }
+
     # term must be a character (not rounded with decimals when integer)
     out$term <- as.character(out$term)
-
 
     if (inherits(out, "data.frame")) {
         return(out)
@@ -154,7 +159,7 @@ These errors messages were generated during extraction:
 
 get_estimates_broom <- function(model, conf_int, conf_level, ...) {
 
-    if (isTRUE(conf_int)) {
+    if (isTRUE(conf_int) && !is.null(conf_level)) {
         out <- suppressWarnings(try(
             broom::tidy(model, conf.int = conf_int, conf.level = conf_level, ...),
             silent = TRUE))
@@ -198,7 +203,7 @@ get_estimates_parameters <- function(model,
     if (isTRUE(conf_int)) {
         args[["ci"]] <- conf_level
     } else if (!"ci" %in% names(dots)) { # faster
-        args <- c(args, list(ci = NULL))
+        args <- c(args, list(ci = NULL, ci_random = FALSE))
     }
 
     # bayes: diagnostics can be very expensive
@@ -207,23 +212,37 @@ get_estimates_parameters <- function(model,
         if (!"diagnostic" %in% names(dots)) args <- c(args, list("diagnostic" = NULL))
     }
 
-    # mixed-effects: ci_random can be very slow in some models
-    if (isTRUE(mi[["is_mixed"]]) &&
-        isTRUE(conf_int) &&
-        !"ci_random" %in% names(args) &&
-        utils::packageVersion("parameters") > "0.18.1.5") {
-        args[["ci_random"]] <- FALSE
+    # mixed-effects: ci_random can be very slow in some models in parameters<=0.18.1
+    fmi <- isTRUE(mi[["is_mixed"]])
+    fci <- isTRUE(conf_int)
+    fra <- !"ci_random" %in% names(args)
+    fve <- utils::packageVersion("parameters") <= "0.18.1"
+    if (fmi && fci && fra && fve) {
+        msg <- format_msg(
+        'Computing confidence intervals for mixed-effects models requires installing a
+        version of the {parameters} package greater than 0.18.1. This version may be
+        available on CRAN:
+
+        install.packages("parameters")
+
+        If {parameters} 0.18.2 is not yet available on CRAN, you can install the
+        development version:
+
+        library(remotes)
+        install_github("easystats/parameters") 
+
+        Make sure you restart R after updating.
+        ')
+        stop(msg, call. = FALSE)
     }
 
     # main call
-    f <- tidy_easystats <- function(x, ...) {
+    fun <- tidy_easystats <- function(x, ...) {
         out <- parameters::parameters(x, ...)
         out <- parameters::standardize_names(out, style = "broom")
     }
 
-    out <- suppressMessages(suppressWarnings(try(
-        do.call("f", args),
-        silent = TRUE)))
+    out <- hush(tryCatch(do.call("fun", args), error = function(e) NULL))
 
     # errors and warnings: before processing the data frame term names
     if (!inherits(out, "data.frame") || nrow(out) < 1) {
@@ -261,3 +280,5 @@ get_estimates_parameters <- function(model,
 
     return(out)
 }
+
+