+++ R/modelsummary.R
@@ -71,7 +71,9 @@ globalVariables(c('.', 'term', 'part', 'estimate', 'conf.high', 'conf.low',
 #' * formula or (named) list of formulas with the cluster variable(s) on the right-hand side (e.g., ~clusterid).
 #' * named list of `length(models)` variance-covariance matrices with row and column names equal to the names of your coefficient estimates.
 #' * a named list of length(models) vectors with names equal to the names of your coefficient estimates. See 'Examples' section below. Warning: since this list of vectors can include arbitrary strings or numbers, `modelsummary` cannot automatically calculate p values. The `stars` argument may thus use incorrect significance thresholds when `vcov` is a list of vectors.
-#' @param conf_level confidence level to use for confidence intervals
+#' @param conf_level numeric value between 0 and 1. confidence level to use for
+#' confidence intervals. Setting this argument to `NULL` does not extract
+#' confidence intervals, which can be faster for some models.
 #' @param exponentiate TRUE, FALSE, or logical vector of length equal to the
 #' number of models. If TRUE, the `estimate`, `conf.low`, and `conf.high`
 #' statistics are exponentiated, and the `std.error` is transformed to
@@ -113,13 +115,16 @@ globalVariables(c('.', 'term', 'part', 'estimate', 'conf.high', 'conf.low',
 #' on columns. The formula can include a group identifier to display related terms
 #' together, which can be useful for models with multivariate outcomes or
 #' grouped coefficients (See examples section below). This identifier must be
-#' one of the column names produced by: `get_estimates(model)`. If an
-#' incomplete formula is supplied (e.g., `~statistic`), `modelsummary` tries to
-#' complete it automatically. Potential `shape` values include:
+#' one of the column names produced by: `get_estimates(model)`. The group
+#' identifier can be combined with the term identifier in a single column by
+#' using the colon to represent an interaction. If an incomplete formula is
+#' supplied (e.g., `~statistic`), `modelsummary` tries to complete it
+#' automatically. Potential `shape` values include:
 #' * `term + statistic ~ model`: default
 #' * `term ~ model + statistic`: statistics in separate columns
 #' * `model + statistic ~ term`: models in rows and terms in columns
-#' * `term + response + statistic ~ model`: 
+#' * `term + response + statistic ~ model`: term and group id in separate columns
+#' * `term : response + statistic ~ model`: term and group id in a single column
 #' * `term ~ response`
 #' @param add_rows a data.frame (or tibble) with the same number of columns as
 #' your main table. By default, rows are appended to the bottom of the table.
@@ -205,11 +210,11 @@ modelsummary <- function(
   estimate <- sanitize_estimate(estimate, number_of_models)
   exponentiate <- sanitize_exponentiate(exponentiate, number_of_models)
   shape <- sanitize_shape(shape)
-  statistic <- sanitize_statistic(statistic, shape) # after shape
+  statistic <- sanitize_statistic(statistic, shape, conf_level) # after shape
   gof_map <- sanitize_gof_map(gof_map)
   fmt <- sanitize_fmt(fmt, calling_function = "modelsummary")
   sanity_group_map(group_map)
-  sanity_conf_level(conf_level)
+  conf_level <- sanitize_conf_level(conf_level, estimate, statistic)
   sanity_coef(coef_map, coef_rename, coef_omit)
   sanity_stars(stars)
   sanity_align(align, estimate = estimate, statistic = statistic, stars = stars)
@@ -238,6 +243,7 @@ modelsummary <- function(
                                         conf_level = conf_level,
                                         vcov = vcov,
                                         gof_map = gof_map, # check if we can skip all gof computation
+                                        shape = shape,
                                         ...)
   names(msl) <- model_names
 
@@ -302,16 +308,18 @@ modelsummary <- function(
       candidate_groups <- unlist(candidate_groups)
       candidate_groups <- setdiff(
         candidate_groups,
-        c("term", "estimate", "std.error", "conf.level", "conf.low", "conf.high",
+        c("term", "type", "estimate", "std.error", "conf.level", "conf.low", "conf.high",
           "statistic", "df.error", "p.value"))
-      if (length(candidate_groups) > 0) {
-        candidate_groups <- sprintf("Candidate group identifiers include: %s.",
-                                    paste(candidate_groups, collapse = ", "))
-      } else{
-        candidate_groups <- ""
-      }
-      msg <- sprintf("There are duplicate term names in the table. The `shape` argument of the `modelsummary` function can be used to print related terms together, and to label them appropriately. You can find the group identifier to use in the `shape` argument by calling `get_estimates()` on one of your models. %s See `?modelsummary` for details.", candidate_groups)
+      msg <- format_msg(
+      "There are duplicate term names in the table.
 
+      The `shape` argument of the `modelsummary` function can be used to print
+      related terms together. The `group_map` argument can be used to reorder,
+      subset, and rename group identifiers. See `?modelsummary` for details.
+
+      You can find the group identifier to use in the `shape` argument by calling
+      `get_estimates()` on one of your models. Candidates include: %s ")
+      msg <- sprintf(msg, paste(candidate_groups, collapse = ", "))
       warning(msg, call. = FALSE)
     }
   }
@@ -509,7 +517,7 @@ modelsummary <- function(
 }
 
 
-get_list_of_modelsummary_lists <- function(models, conf_level, vcov, gof_map, ...) {
+get_list_of_modelsummary_lists <- function(models, conf_level, vcov, gof_map, shape, ...) {
 
     number_of_models <- max(length(models), length(vcov))
 
@@ -544,25 +552,40 @@ get_list_of_modelsummary_lists <- function(models, conf_level, vcov, gof_map, ..
         # recycling when 1 model and many vcov
         j <- ifelse(length(models) == 1, 1, i)
 
+        if (inherits(models[[j]], "modelsummary_list")) {
+            out <- list(
+                tidy = models[[j]][["tidy"]],
+                glance = models[[j]][["glance"]])
+            return(out)
+        }
+
         # don't waste time if we are going to exclude all gof anyway
         gla <- get_gof(models[[j]], vcov_type[[i]], gof_map = gof_map, ...)
 
-        tid <- get_estimates(models[[j]], conf_level = conf_level, vcov = vcov[[i]], ...)
+        tid <- get_estimates(
+            models[[j]],
+            conf_level = conf_level,
+            vcov = vcov[[i]],
+            shape = shape,
+            ...)
 
         out <- list("tidy" = tid, "glance" = gla)
         class(out) <- "modelsummary_list"
         return(out)
     }
 
-    # parallel
-    if (isTRUE(check_dependency("future.apply"))) {
-        parallel_flag <- !"sequential" %in% attr(future::plan(), "class") && number_of_models > 1
-    } else {
-        parallel_flag <- FALSE
-    }
+    # {parallel}
+    dots <- list(...)
+    if ("mc.cores" %in% names(dots)) {
+        out <- parallel::mclapply(seq_len(number_of_models), inner_loop, mc.cores = dots[["mc.cores"]])
 
-    if (isTRUE(parallel_flag)) {
-        out <- future.apply::future_lapply(seq_len(number_of_models), inner_loop)
+    # {future}
+    } else if (isTRUE(check_dependency("future.apply")) &&
+               future::nbrOfWorkers() > 1 &&
+               number_of_models > 1) {
+        out <- future.apply::future_lapply(seq_len(number_of_models), inner_loop, future.seed = TRUE)
+
+    # sequential
     } else {
         out <- lapply(seq_len(number_of_models), inner_loop)
     }
@@ -590,3 +613,6 @@ redundant_labels <- function(dat, column) {
 #' @keywords internal
 #' @export
 msummary <- modelsummary
+
+
+